requirements.txt
flask
requests
pymongo
twilio
pandas
openpyxl
python-dotenv